requirements.txt
asgiref==3.7.2
astroid==3.0.1
dill==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.1.0
gunicorn==21.2.0
isort==5.12.0
mccabe==0.7.0
packaging==23.2
Pillow==10.1.0
platformdirs==4.0.0
psycopg2-binary==2.9.9
pylint==3.0.2
pylint-celery==0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
sqlparse==0.4.4
tomlkit==0.12.3
typing_extensions==4.8.0
